new ipfailoversckript, changed kopano to supported version

+++ kopano/entrypoint.sh
@@ -24,10 +24,15 @@ sed -i "s/^#log_level.*/log_level = $KOPANO_SERVER_LOGLEVEL/g" /etc/kopano/spool
 sed -i "s/^smtp_server.*/smtp_server = $SMTP_SERVER/g" /etc/kopano/spooler.cfg
 
 
-
 cp /tmp/template/kopano/20-kopano.ini /etc/php/7.0/fpm/conf.d/20-kopano.ini
-cp /tmp/template/services/* /etc/init.d/
-chmod 777 /etc/init.d/kopano-*
+
+#Kopieren von Kopano Serice Start Skripten
+if [ ! -f "/etc/init.d/kopano-server"  ]; then
+    cp /tmp/template/services/* /etc/init.d/ 
+	chmod 777 /etc/init.d/kopano-*
+	echo "Startscripts created"
+fi
+
 
 #Anpassen der Nginx Config
 cp /tmp/template/nginx/nginx.conf /etc/nginx/nginx.conf
@@ -51,6 +56,20 @@ echo "php_flag register_globals = off" >> /etc/php/7.0/fpm/php.ini
 echo "php_flag magic_quotes_runtime = off" >> /etc/php/7.0/fpm/php.ini
 echo "php_flag short_open_tag = on" >> /etc/php/7.0/fpm/php.ini
 
+#Anpassen der PHP-FPM Config
+sed -i "s/;pm.status_path = \/status/pm.status_path = \/status/g" /etc/php/7.0/fpm/pool.d/www.conf
+sed -i "s/pm.max_children.*/pm.max_children = 30/g" /etc/php/7.0/fpm/pool.d/www.conf
+sed -i "s/pm.start_servers.*/pm.start_servers = 5/g" /etc/php/7.0/fpm/pool.d/www.conf
+sed -i "s/pm.min_spare_servers.*/pm.min_spare_servers = 5/g" /etc/php/7.0/fpm/pool.d/www.conf
+sed -i "s/pm.max_spare_servers.*/pm.max_spare_servers = 10/g" /etc/php/7.0/fpm/pool.d/www.conf
+sed -i "s/;pm.max_requests.*/pm.max_requests = 500/g" /etc/php/7.0/fpm/pool.d/www.conf
+sed -i "s/;request_slowlog_timeout.*/request_slowlog_timeout = 5s/g" /etc/php/7.0/fpm/pool.d/www.conf
+sed -i "s/;slowlog.*/slowlog = \/var\/log\/slowlog.log.slow/g" /etc/php/7.0/fpm/pool.d/www.conf
+
+sed -i "s/;emergency_restart_threshold.*/emergency_restart_threshold = 10/g" /etc/php/7.0/fpm/php-fpm.conf
+sed -i "s/;emergency_restart_interval.*/emergency_restart_interval = 1m/g" /etc/php/7.0/fpm/php-fpm.conf
+sed -i "s/;process_control_timeout.*/process_control_timeout = 10/g" /etc/php/7.0/fpm/php-fpm.conf
+
 #Starten von Nginx
 service nginx start
 
@@ -58,18 +77,15 @@ service nginx start
 if [ ! -z "$NGINX_API_KEY" ]; then
 	echo "Start Nginx Amplify"
 	
-	#Anpassen der PHP-FPM Config
-	sed -i "s/;pm.status_path = \/status/pm.status_path = \/status/g" /etc/php/7.0/fpm/pool.d/www.conf 
-	
 	#Anpassen der Amplify Config
 	sed -i "s/api_key .*/api_key = $NGINX_API_KEY/g" /etc/amplify-agent/agent.conf
 	sed -i "s/hostname .*/hostname = kopano\.$DOMAIN/g" /etc/amplify-agent/agent.conf
+	sed -i "s/uuid =.*/uuid = 0f2c2dcf00205d45a6eba415b862bba1/g" /etc/amplify-agent/agent.conf
 	
 	#Starten des Amplify Dienstes
 	service amplify-agent start
 else
 	echo "Nginx Amplify wird nicht gestartet ... NGINX_API_KEY wurde nicht übergeben"
-
 fi
 
 #Starten der Services
